Replace AssetDurationLoader with an AVURLAsset extension
Load media duration through AVURLAsset.loadDuration() instead of a
separate helper type.
